Add factory method for favorites.
Makes creating the favorite a bit more reusable and avoids duplicating
logic to builders.
