+++ opengever/base/model/favorite.py
@@ -1,3 +1,4 @@
+from opengever.base.browser.helper import get_css_class
 from opengever.base.date_time import utcnow_tz_aware
 from opengever.base.model import Base
 from opengever.base.model import CONTENT_TITLE_LENGTH
@@ -16,6 +17,7 @@ from opengever.dossier.dossiertemplate.behaviors import IDossierTemplateMarker
 from opengever.ogds.models.admin_unit import AdminUnit
 from opengever.repository.interfaces import IRepositoryFolder
 from plone import api
+from plone.uuid.interfaces import IUUID
 from Products.CMFPlone.utils import safe_unicode
 from sqlalchemy import and_
 from sqlalchemy import Boolean
@@ -61,6 +63,33 @@ class Favorite(Base):
     is_subdossier = Column(Boolean, nullable=True)
     is_leafnode = Column(Boolean, nullable=True)
 
+    @classmethod
+    def create(cls, userid, obj, **kwargs):
+        truncated_title = cls.truncate_title(obj.Title().decode('utf-8'))
+
+        is_subdossier = None
+        if IDossierMarker.providedBy(obj) or IDossierTemplateMarker.providedBy(obj):
+            is_subdossier = obj.is_subdossier()
+
+        is_leafnode = None
+        if IRepositoryFolder.providedBy(obj):
+            is_leafnode = obj.is_leaf_node()
+
+        params = dict(
+            userid=userid,
+            oguid=Oguid.for_object(obj),
+            title=truncated_title,
+            portal_type=obj.portal_type,
+            icon_class=get_css_class(obj),
+            plone_uid=IUUID(obj),
+            position=cls.query.get_next_position(userid),
+            review_state=api.content.get_state(obj),
+            is_subdossier=is_subdossier,
+            is_leafnode=is_leafnode,
+        )
+        params.update(kwargs)
+        return cls(**params)
+
     def serialize(self, portal_url):
         return {
             '@id': self.api_url(portal_url),
@@ -105,6 +134,13 @@ class Favorite(Base):
 
 
 class FavoriteQuery(BaseQuery):
+
+    def get_next_position(self, userid):
+        position = self.get_highest_position(userid)
+        if position is None:
+            return 0
+
+        return position + 1
 
     def is_marked_as_favorite(self, obj, user):
         favorite = self.by_object_and_user(obj, user).first()
